feat: sa-251 - refactored the handler and renders - makes testing easier
